Renamed folder and namespace to match the ones for the other wizards.
