+++ XenAdmin/Commands/BugToolCommand.cs
@@ -29,7 +29,7 @@
  */
 
 using XenAdmin.Network;
-using XenAdmin.Wizards.BugToolWizardFiles;
+using XenAdmin.Wizards.BugToolWizard;
 using XenAPI;
 
 namespace XenAdmin.Commands
