+++ XenAdmin/Wizards/BugToolWizard/BugToolPageSelectCapabilities.ja.resx
