+++ XenAdmin/Wizards/BugToolWizard/BugToolPageSelectCapabilities.resx
