+++ XenAdmin/Wizards/BugToolWizard/BugToolPageSelectCapabilities.zh-CN.resx
